requirements.txt
# Frontend UI
streamlit

# PDF parsing
pymupdf

# Environment variables
python-dotenv